Split first run panels into separate views
