+++ xcode/Subconscious/Shared/Components/FirstRun/FirstRunView.swift
@@ -90,7 +90,7 @@ struct FirstRunView: View {
                 VStack(alignment: .leading, spacing: AppTheme.unit3) {
                     Text("Welcome to Subconscious, a place to garden thoughts and share them with others.")
                     
-                    Text("Subconscious is powered by decentralized protocols that put you in control. Your thoughts belong to you forever.")
+                    Text("Subconscious is powered by an open and decentralized note graph. Your data is yours, forever.")
                 }
                 .foregroundColor(.secondary)
                 .font(.callout)
@@ -112,72 +112,6 @@ struct FirstRunView: View {
         .background(.background)
         .task {
             store.send(.start)
-        }
-    }
-}
-
-struct FirstRunCreateSphereView: View {
-    /// FirstRunView is a major view that manages its own state in a store.
-    @ObservedObject var store: Store<FirstRunModel>
-    var done: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            VStack {
-                Spacer()
-                VStack(alignment: .leading) {
-                    Text("Recovery Passphrase")
-                        .font(.headline)
-                    HStack {
-                        Text(store.state.sphereMnemonic ?? "")
-                            .monospaced()
-                        Spacer()
-                    }
-                    .padding()
-                    .overlay(
-                        RoundedRectangle(cornerRadius: AppTheme.cornerRadiusLg)
-                            .stroke(Color.separator, lineWidth: 0.5)
-                    )
-                }
-                Text("This is your notebook's recovery passphrase. It's for you alone. We don't store it. Write it down someplace, keep it safe.")
-                    .foregroundColor(.secondary)
-                Spacer()
-                NavigationLink(
-                    destination: {
-                        FirstRunDone(store: store, done: done)
-                    },
-                    label: {
-                        Text("Continue")
-                    }
-                )
-            }
-            .padding()
-        }
-    }
-}
-
-struct FirstRunDone: View {
-    /// FirstRunView is a major view that manages its own state in a store.
-    @ObservedObject var store: Store<FirstRunModel>
-    var done: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            VStack {
-                Spacer()
-                VStack(spacing: AppTheme.unit3) {
-                    Image(systemName: "checkmark.circle.fill")
-                        .font(.system(size: 96))
-                        .foregroundColor(.accentColor)
-                    Text("All set!")
-                        .foregroundColor(.secondary)
-                }
-                Spacer()
-                Button(action: done) {
-                    Text("Continue")
-                }
-            }
-            .padding()
         }
     }
 }
